Fix timestamp precision loss on MySQL; connect to real self-hosted DB
deadline/remaining_seconds/created_at used SQLAlchemy's generic Float,
which compiles to MySQL's 32-bit single-precision FLOAT (~7 significant
digits) -- not enough for a ~10-digit Unix timestamp. A study_seconds=600
session's deadline came back several hundred seconds into the past on
read-back. SQLite (what tests use) stores Python floats at full 64-bit
precision, so this was invisible until tested against real MySQL.

Fixed by switching every timestamp/deadline column to SQLAlchemy's
Double type (DOUBLE on MySQL, unaffected on SQLite). The four 0-100-range
WellbeingReading metrics stayed Float -- float32 precision is plenty for
that range, only large-magnitude values like timestamps were affected.

Verified against the team's actual database (self-hosted MySQL on a
Vultr VPS, not the Managed Database product) via a live end-to-end run
(signup -> upload -> session -> chat -> wellbeing -> recall -> advance)
against the real Gemini API: deadline math is exact post-fix. Docs
updated with the real connection steps taken and this bug's writeup.

+++ backend/models/document.py
@@ -1,5 +1,5 @@
 import time
-from sqlalchemy import Column, Integer, String, Text, JSON, LargeBinary, ForeignKey, ForeignKeyConstraint, Float
+from sqlalchemy import Column, Integer, String, Text, JSON, LargeBinary, ForeignKey, ForeignKeyConstraint, Double
 from backend.db.base_class import Base
 
 class Document(Base):
@@ -20,7 +20,9 @@ class Document(Base):
     original = Column(LargeBinary, nullable=False)
     progress = Column(JSON, nullable=False, default=dict)
     revision = Column(Integer, nullable=False, default=1)
-    created_at = Column(Float, default=time.time, nullable=False)
+    # Double, not Float -- see the note on StudySession.deadline in
+    # models/study.py for why a Unix timestamp needs it.
+    created_at = Column(Double, default=time.time, nullable=False)
     __mapper_args__ = {"version_id_col": revision}
 
 
@@ -46,7 +48,7 @@ class DocumentObject(Base):
     # (document_service.check_pages) needs all of them. Column NAME kept as
     # `source_page` to match the schema; VALUE is a list[int].
     source_page = Column(JSON, nullable=True)
-    created_at = Column(Float, default=time.time, nullable=False)
+    created_at = Column(Double, default=time.time, nullable=False)  # see the Double note above
     __table_args__ = (
         ForeignKeyConstraint(
             ["uid", "document_id"], ["documents.uid", "documents.document_id"],
